forecast-model-trainer/pyproject.toml
[tool.poetry]
name = "forecast-model-trainer"
version = "0.1.0"
description = "Project for training weather forecast model."
authors = ["Jan Breznar, Janez"]
readme = "README.md"
packages = [{include = "forecast_model_trainer"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.92.0"
pandas = "^1.5.3"
numpy = "=1.23.5"
scikit-learn = "^1.2.1"
pytest = "^7.2.1"
uvicorn = "^0.20.0"
httpx = "^0.23.3"
great-expectations = "^0.16.3"
evidently = "^0.2.7"
mlflow = "^2.2.2"
pymongo = {extras = ["srv"], version = "^4.3.3"}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
